pyproject.toml
[project]
name = "jedi-mcp"
version = "0.1.0"
description = "Convert technical documentation websites into Model Context Protocol (MCP) servers"
authors = [
    {name = "Jedi-MCP Contributors"}
]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "strands-agents>=0.1.0",
    "fastmcp>=0.1.0",
    "httpx>=0.24.0",
    "beautifulsoup4>=4.12.0",
    "click>=8.1.0",
    "lxml>=4.9.0",
    "python-dotenv>=1.0.0",
    "playwright>=1.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "hypothesis>=6.82.0",
    "responses>=0.23.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
jedi-mcp = "jedi_mcp.cli:main"

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
